requirements.txt
django==3.0.3
django-guardian==2.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
psycopg2-binary==2.8.4
Pillow==7.0.0
